requirements.txt
Django==3.2.8
psycopg2
djangorestframework==3.12.4
pytest-django
pytest